Sort anchors by their associatedPageIndex value
This value is always integer, so we don't have to parse things
like Roman numerals (which can exist in anchor text itself),
and we can impose meaningful order on arbitrary string anchors
(which can be surrounded by numeric anchors).

For example, anchor "pg5" can be followed by "pgimage1", then by "pg6".

We are intentionally ignoring anchors with negative numbers,
as this is a convention for anchors that are not supposed to be
in the navigation template.